requirements.txt
# ---- Web Interface Core ----
# Used to build the frontend website UI simply using Python.
streamlit

# ---- Environment & Data Utilities ----
# Used to safely load hidden API keys from the .env file.
python-dotenv
# Used for downloading webpages and interacting with external APIs.
requests
# Used for reading the messy HTML code of webpages to extract clean text.
beautifulsoup4
# Used for manipulating and displaying images (like the robot icon).
pillow

# ---- Core AI & Agent Framework ----
# LangChain is the main framework used to connect to AI models and give them tools.
langchain
langchain-core
langchain-community
langchain-text-splitters
# LangGraph is the advanced brain used to loop AI thoughts (Agents) and route them.
langgraph
langgraph-prebuilt

# ---- Language Models ----
# Connects specifically to Google's Gemini Models
langchain-google-genai
# Connects specifically to OpenAI's GPT Models
langchain-openai

# ---- Vector Database for Document Processing (RAG) ----
# Chroma is the database used to store vectorized text chunks from PDFs.
chromadb
# Used to convert text chunks into numbers (embeddings) for free.
sentence-transformers
# Used to read the text content out of PDF files.
pypdf

# ---- AI Tools ----
# Grants the AI the ability to search the live internet.
duckduckgo-search
